refactor config loader
