+++ apollo-client/src/main/java/com/ctrip/apollo/client/ApolloConfig.java
@@ -1,6 +1,7 @@
 package com.ctrip.apollo.client;
 
-import com.ctrip.apollo.client.impl.MockConfigLoader;
+import com.ctrip.apollo.client.loader.ConfigLoader;
+import com.ctrip.apollo.client.loader.ConfigLoaderFactory;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
@@ -26,7 +27,7 @@ public class ApolloConfig implements BeanDefinitionRegistryPostProcessor, Priori
     private ConfigurableApplicationContext applicationContext;
 
     public ApolloConfig() {
-        this.configLoader = new MockConfigLoader();
+        this.configLoader = ConfigLoaderFactory.getInstance().getMockConfigLoader();
     }
 
     @Override
