+++ apollo-client/src/main/java/com/ctrip/apollo/client/loader/ConfigLoaderFactory.java
@@ -1,0 +1,21 @@
+package com.ctrip.apollo.client.loader;
+
+import com.ctrip.apollo.client.loader.impl.MockConfigLoader;
+
+/**
+ * @author Jason Song(song_s@ctrip.com)
+ */
+public class ConfigLoaderFactory {
+    private static ConfigLoaderFactory configLoaderFactory = new ConfigLoaderFactory();
+
+    private ConfigLoaderFactory() {
+    }
+
+    public static ConfigLoaderFactory getInstance() {
+        return configLoaderFactory;
+    }
+
+    public ConfigLoader getMockConfigLoader() {
+        return new MockConfigLoader();
+    }
+}
